# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"@project.name@","description":"@project.description@","contact":{"name":"None","email":"none@none.com"},"version":"@project.version@"},"servers":[{"url":"http://localhost:8080","description":"@project.description@"}],"paths":{"/updateRent":{"put":{"tags":["rent-controller"],"operationId":"putRent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRentDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RentResponseDTO"}}}}}}},"/updateCar":{"put":{"tags":["car-controller"],"operationId":"putCar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCarDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarResponseDTO"}}}}}}},"/newrent":{"post":{"tags":["rent-controller"],"operationId":"postRent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RentResponseDTO"}}}}}}},"/newType":{"post":{"tags":["car-type-controller"],"operationId":"postType","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarTypeResponseDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarTypeResponseDTO"}}}}}}},"/newCar":{"post":{"tags":["car-controller"],"operationId":"postCar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarResponseDTO"}}}}}}},"/rentsAvailabilityByDate/{date}":{"get":{"tags":["rent-controller"],"operationId":"getCarAvailabilityByDate","parameters":[{"name":"date","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CarResponseDTO"}}}}}}}},"/rent/{id}":{"get":{"tags":["rent-controller"],"operationId":"getRentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RentResponseDTO"}}}}}}},"/hola":{"get":{"tags":["hola-controller"],"operationId":"hola","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/cars":{"get":{"tags":["car-controller"],"operationId":"getAllCars","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CarResponseDTO"}}}}}}}},"/carsEntity":{"get":{"tags":["car-controller"],"operationId":"getAllCarsEntity","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CarEntity"}}}}}}}},"/carTypes":{"get":{"tags":["car-type-controller"],"operationId":"getAllTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CarTypeResponseDTO"}}}}}}}},"/car/{id}":{"get":{"tags":["car-controller"],"operationId":"getCarById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarResponseDTO"}}}}}}},"/allRents":{"get":{"tags":["rent-controller"],"operationId":"getAllRents","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RentResponseDTO"}}}}}}}},"/deleteRent/{id}":{"delete":{"tags":["rent-controller"],"operationId":"deleteRent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/deleteCar/{id}":{"delete":{"tags":["car-controller"],"operationId":"deleteCar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarResponseDTO"}}}}}}}},"components":{"schemas":{"UpdateRentDTO":{"required":["carId","id"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"customer":{"type":"string"},"carId":{"type":"integer","format":"int64"},"startRent":{"type":"string","format":"date-time"},"endRent":{"type":"string","format":"date-time"}}},"RentResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"customer":{"type":"string"},"carType":{"type":"string"},"carBrand":{"type":"string"},"carModel":{"type":"string"},"rentedDays":{"type":"integer","format":"int32"},"startRent":{"type":"string","format":"date-time"},"endRent":{"type":"string","format":"date-time"},"totalPrice":{"type":"number"}}},"UpdateCarDTO":{"required":["id","model"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"carTypeId":{"type":"integer","format":"int64"},"brand":{"type":"string"},"model":{"type":"string"}}},"CarResponseDTO":{"required":["brand","model","type"],"type":"object","properties":{"type":{"type":"string"},"brand":{"type":"string"},"model":{"type":"string"}}},"RentRequestDTO":{"required":["carId","endRent","startRent"],"type":"object","properties":{"customer":{"type":"string"},"carId":{"type":"integer","format":"int64"},"startRent":{"type":"string","format":"date-time"},"endRent":{"type":"string","format":"date-time"}}},"CarTypeResponseDTO":{"required":["id","type"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"}}},"CarRequestDTO":{"required":["brand","carType","model"],"type":"object","properties":{"carType":{"type":"integer","format":"int64"},"brand":{"type":"string"},"model":{"type":"string"}}},"CarEntity":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"typeId":{"$ref":"#/components/schemas/CarTypeEntity"},"brand":{"type":"string"},"model":{"type":"string"}}},"CarTypeEntity":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"price":{"type":"number"}}}}}}