# src/main/java/ar/edu/utn/frc/tup/lciii/finalpractica/Entities/CarEntity.java
package ar.edu.utn.frc.tup.lciii.finalpractica.Entities;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity
@Table(name="cars")
public class CarEntity {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne
    @JoinColumn(name="car_type_id")
    private CarTypeEntity typeId;
    @Column
    private String brand;
    @Column
    private String model;
}