# src/main/java/ar/edu/utn/frc/tup/lciii/finalpractica/Entities/CarTypeEntity.java
package ar.edu.utn.frc.tup.lciii.finalpractica.Entities;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity
@Table(name = "cars_type")
public class CarTypeEntity {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    @Column(name="type")
    private String type;
    @Column(name="price")
    private BigDecimal price;
}